# Make `status` significantly faster